aipart/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-genai==0.2.2
pillow==10.1.0
python-dotenv==1.0.0
stripe==13.0.1
resend==2.17.0
reportlab==4.4.4
gunicorn==21.2.0
